se cambia la carpeta de los archivos siendo esta a privado y se modifican imports
