+++ EduConnect-FrontEnd/src/App.jsx
@@ -2,9 +2,9 @@ import { Routes, Route } from 'react-router-dom'
 import PerfilSuperAdminFormulario from './views/PerfilSuperAdminFormulario.jsx'
 import PerfilSuperAdminHome from './views/PerfilSuperAdminHome.jsx'
 import PerfilSuperAdminFormularioAdmin from './views/PerfilSuperAdminFormularioAdmin.jsx'
-import PerfilAdminFormularioProfesor from './views/PerfilAdminFormularioProfesor.jsx'
+import PerfilAdminFormularioProfesor from './views/privado/PerfilAdminFormularioProfesor.jsx'
 import PerfilAdminHome from './views/PerfilAdminHome.jsx'
-import PerfilAdminFormularioEstudiante from './views/PerfilAdminFormularioEstudiante.jsx'
+import PerfilAdminFormularioEstudiante from './views/privado/PerfilAdminFormularioEstudiante.jsx'
 import NavBar from './components/publico/NavBar.jsx'
 import HomePage from './views/publico/HomePage.jsx'
 import Login from './views/publico/login.jsx'
